feat: move testUtils to shared utils folder
