+++ apps/server/src/utils/repeatActivity/repeatActivityThisMonth/repeatActivityThisMonth.spec.ts
@@ -3,7 +3,7 @@ import { WEEK_OF_MONTH } from "../../mocks";
 import {
   createMockActivityByDateString,
   createMockWeekFormDateStringArray,
-} from "../testUtils";
+} from "../../testUtils";
 import { repeatActivityThisMonth } from "./repeatActivityThisMonth";
 
 describe("repeatActivityThisMonth", () => {
