+++ apps/server/src/utils/testUtils.ts
@@ -1,6 +1,7 @@
 import { Activity, RepeatTypeEnum } from "@routine-support/domains";
 import { DateString } from "@routine-support/types";
 import { parseDate } from "@routine-support/utils";
+import moment from "moment";
 import { Moment } from "moment";
 
 export const createMockActivityByDateString = (date: DateString): Activity => {
@@ -13,7 +14,22 @@ export const createMockActivityByDateString = (date: DateString): Activity => {
     start: parsedDate,
     end: parsedDate,
     name: "",
-    repeat: RepeatTypeEnum.None,
+    repeatType: RepeatTypeEnum.None,
+    confirmation: {},
+  };
+};
+
+export const createMockActivityByRepeatType = (repeatType: RepeatTypeEnum) => {
+  const mockedDate = moment();
+
+  return {
+    date: mockedDate,
+    coachId: "",
+    pictogram: "",
+    start: mockedDate,
+    end: mockedDate,
+    name: "",
+    repeatType,
     confirmation: {},
   };
 };
